web_automation/requirements.txt
easyocr==1.7.2
numpy==2.2.6
pillow==12.0.0
pyppeteer==2.0.0
torch==2.9.1
opencv-python-headless==4.12.0.88

